+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [12.8.0]
+- [iOS] Fixed an issue where touch effect would still be imminent when a user taps and then slides the finger out of bounds
+- [Android] Fixed an issue where touch effect would go out of bounds on some elements
+
 ## [12.7.0]
 - [Android] Added a workaround to enable toolbar for modal pages